scripts/requirements.txt
psycopg2-binary==2.9.9
httpx==0.27.0
pandas==2.2.2
pdfplumber==0.11.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.1